Dealias opaque types to their translucent super type
